move FilteredResource and make include/opt-in instead of omit
